Add files via upload
The dataset we're working with is called "spa-eng"
"http://storage.googleapis.com/download.tensorflow.org/data/spa-eng.zip"
and it contains parallel text data in Spanish and English languages. Each entry in the dataset consists of a pair of sentences: one in Spanish and its corresponding translation in English. This dataset is commonly used for tasks such as machine translation, where the goal is to translate text from one language to another.
we are utilizing this dataset to train and evaluate a neural machine translation model. The model learns from the parallel sentences in Spanish and English to understand the relationship between words and phrases in both languages, enabling it to generate accurate translations for new input sentences.
Using this dataset, our goal is to create a strong and reliable machine translation system that can accurately translate languages in real-world scenarios.